+++ crates/teloxide-core/src/requests.rs
@@ -20,4 +20,3 @@ mod payload;
 mod request;
 mod requester;
 mod requester_ext;
-mod utils;