+++ crates/teloxide/src/dispatching/dialogue/get_chat_id.rs
@@ -1,6 +1,8 @@
-use crate::types::{CallbackQuery, ChatId, Message, Update};
+use crate::types::{
+    CallbackQuery, Chat, ChatId, ChatJoinRequest, ChatMemberUpdated, Message, Update,
+};
 
-/// Something that may has a chat ID.
+/// Something that may have a chat ID.
 pub trait GetChatId {
     #[must_use]
     fn chat_id(&self) -> Option<ChatId>;
@@ -21,5 +23,23 @@ impl GetChatId for CallbackQuery {
 impl GetChatId for Update {
     fn chat_id(&self) -> Option<ChatId> {
         self.chat().map(|chat| chat.id)
+    }
+}
+
+impl GetChatId for Chat {
+    fn chat_id(&self) -> Option<ChatId> {
+        Some(self.id)
+    }
+}
+
+impl GetChatId for ChatMemberUpdated {
+    fn chat_id(&self) -> Option<ChatId> {
+        Some(self.chat.id)
+    }
+}
+
+impl GetChatId for ChatJoinRequest {
+    fn chat_id(&self) -> Option<ChatId> {
+        Some(self.chat.id)
     }
 }