+++ crates/teloxide/src/dispatching/dialogue/storage.rs
@@ -6,7 +6,7 @@ mod trace_storage;
 #[cfg(feature = "redis-storage")]
 mod redis_storage;
 
-#[cfg(feature = "sqlite-storage-nativetls")]
+#[cfg(any(feature = "sqlite-storage-nativetls", feature = "sqlite-storage-rustls"))]
 mod sqlite_storage;
 
 #[cfg(feature = "postgres-storage-nativetls")]
@@ -25,7 +25,7 @@ pub use redis_storage::{RedisStorage, RedisStorageError};
 pub use serializer::Serializer;
 use std::sync::Arc;
 
-#[cfg(feature = "sqlite-storage-nativetls")]
+#[cfg(any(feature = "sqlite-storage-nativetls", feature = "sqlite-storage-rustls"))]
 pub use sqlite_storage::{SqliteStorage, SqliteStorageError};
 
 #[cfg(feature = "postgres-storage-nativetls")]